+++ src/components/Icon/Icon.tsx
@@ -1,11 +1,12 @@
 import { IconProp } from "@fortawesome/fontawesome-svg-core";
 import {
   FontAwesomeIcon,
-  FontAwesomeIconProps,
+  FontAwesomeIconProps
 } from "@fortawesome/react-fontawesome";
 import cx from "classnames";
 import { forwardRef } from "react";
 import { OmitStrict } from "../../utils/types";
+import { COMPONENT_PREFIX } from "../resources/common.data";
 
 interface IconInfo {
   icon: [string, string];
@@ -16,7 +17,7 @@ export const Icon = forwardRef<
   HTMLDivElement,
   Partial<OmitStrict<FontAwesomeIconProps, "icon" | "cx">> & IconInfo
 >(({ icon, className, ...faProps }, ref) => (
-  <div ref={ref} className={cx("icon", className)}>
+  <div ref={ref} className={cx(`${COMPONENT_PREFIX}-icon`, "icon", className)}>
     <FontAwesomeIcon {...faProps} icon={icon as IconProp} />
   </div>
 ));