+++ src/components/Listbox/AsyncListbox.tsx
@@ -5,7 +5,7 @@ import { OmitStrict } from "../../utils/types";
 import { Icon } from "../Icon";
 import {
   getListboxClassName,
-  listboxClassNamePrefix,
+  listboxInnerClassNamePrefix,
   listboxNoOptionMessage,
   listBoxOnChange,
   ListboxOption,
@@ -57,7 +57,7 @@ export function AsyncListbox<Data, IsMulti extends boolean = false, Form = any>(
             selectProps.isMulti,
             formOnChange,
           )}
-          classNamePrefix={listboxClassNamePrefix}
+          classNamePrefix={listboxInnerClassNamePrefix}
           components={{
             DropdownIndicator: (props) => (
               <Icon