+++ src/components/Listbox/CreatableListbox.tsx
@@ -5,7 +5,7 @@ import { OmitStrict } from "../../utils/types";
 import { Icon } from "../Icon";
 import {
   getListboxClassName,
-  listboxClassNamePrefix,
+  listboxInnerClassNamePrefix,
   listboxNoOptionMessage,
   listBoxOnChange,
   ListboxOption,
@@ -59,7 +59,7 @@ export function CreatableListbox<
             selectProps.isMulti,
             formOnChange,
           )}
-          classNamePrefix={listboxClassNamePrefix}
+          classNamePrefix={listboxInnerClassNamePrefix}
           components={{
             DropdownIndicator: (props) => (
               <Icon