+++ src/utils/date.ts
@@ -1,6 +1,8 @@
 import dayjs from "dayjs";
 
-export const STANDARD_DATE_FORMAT = "DD/MM/YYYY";
+export const STANDARD_TIME_FORMAT = "HH:mm";
+export const STANDARD_DATE_FORMAT = "MM/DD/YYYY";
+export const STANDARD_DATE_TIME_FORMAT = "MM/DD/YYYY HH:mm";
 
 export const formatDate = (
   date: string | Date,
@@ -14,6 +16,9 @@ export const formatDateRange = (
 
 // TODO: Support locale for this
 export const getDatePlaceholder = () => "__/__/____";
+
+// TODO: Support locale for this
+export const getTimePlaceholder = () => "__:__";
 
 /** Sort date array IN PLACE */
 export const sortDateArray = <T>({